Revert "Merge branch 'master' into palette-browser"
This reverts commit a0033d40337c34c2785d7fff2ca29f72cda6a411.

+++ scilab/modules/hdf5/src/cpp/H5BitfieldData.hxx
@@ -85,12 +85,12 @@ public:
 
 #ifdef __SCILAB_INT64__
 
-class H5Bitfield8Data : public H5BasicData<unsigned long long>
+class H5Bitfield8Data : public H5BasicData<unsigned long>
 {
 
 public:
 
-    H5Bitfield8Data(H5Object & _parent, const hsize_t _totalSize, const hsize_t _dataSize, const hsize_t _ndims, const hsize_t * _dims, unsigned long long * _data, const hsize_t _stride, const size_t _offset, const bool _dataOwner) : H5BasicData<unsigned long long>(_parent, _totalSize, _dataSize, _ndims, _dims, _data, _stride, _offset, _dataOwner)
+    H5Bitfield8Data(H5Object & _parent, const hsize_t _totalSize, const hsize_t _dataSize, const hsize_t _ndims, const hsize_t * _dims, unsigned long * _data, const hsize_t _stride, const size_t _offset, const bool _dataOwner) : H5BasicData<unsigned long>(_parent, _totalSize, _dataSize, _ndims, _dims, _data, _stride, _offset, _dataOwner)
     {
 
     }
